feat: not subscribe segment
