+++ lib/config/routes.dart
@@ -17,10 +17,10 @@ import 'package:tawkie/pages/chat_permissions_settings/chat_permissions_settings
 import 'package:tawkie/pages/device_settings/device_settings.dart';
 import 'package:tawkie/pages/invitation_selection/invitation_selection.dart';
 import 'package:tawkie/pages/login/login.dart';
-import 'package:tawkie/pages/login/not_subscribe.dart';
 import 'package:tawkie/pages/new_group/new_group.dart';
 import 'package:tawkie/pages/new_private_chat/new_private_chat.dart';
 import 'package:tawkie/pages/new_space/new_space.dart';
+import 'package:tawkie/pages/not_subscribe/not_subscribe_page.dart';
 import 'package:tawkie/pages/settings/settings.dart';
 import 'package:tawkie/pages/settings_3pid/settings_3pid.dart';
 import 'package:tawkie/pages/settings_chat/settings_chat.dart';
